Present paired live Rust examples through the native renderer
